% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poids_mgpp.R
\name{namedList}
\alias{namedList}
\title{Title namedList}
\usage{
namedList(...)
}
\arguments{
\item{...}{elements de la liste}
}
\value{
liste dont les elements ont les noms des objets composant la liste
}
\description{
Title namedList
}
