% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_indirecte_fast.R
\name{poids_ind}
\alias{poids_ind}
\title{Probabilité d'inclusion simple indirecte (version rapide)}
\usage{
poids_ind(
  UA = NULL,
  dataStrate = NULL,
  l_ji = NULL,
  id_j = "idA",
  id_i = "idB",
  nomStrate = "strate",
  population = "Nh",
  echantillon = "nh",
  controles = T
)
}
\arguments{
\item{UA}{base de sondage (variables id_j + aux_j)}

\item{dataStrate}{dataframe decrivant les strates}

\item{l_ji}{table des liens (variables id_j, id_i)}

\item{id_j}{identifiant au niveau UA}

\item{id_i}{identifiant au niveau UB}

\item{nomStrate}{variable strate dans UA et dans dataStrate}

\item{population}{variable indiquant le nombre d'unite dans la BDS dans la strate dans dataStrate}

\item{echantillon}{variable indiquant le nombre d'unite a tirer dans la strate dans dataStrate}

\item{controles}{effectuer des controles sur les tables (default=T)}
}
\value{
dataframe avec 2 colonnes :
\itemize{
  \item id_i : identifiant au niveau UB
  \item w_ind : poids correspondant à l'inverse de la probabilité d'inclusion
}
}
\description{
Cette fonction est une version de \code{\link{proba_inclusion_indirecte}} avec un format proche
de \code{\link{poids_mgpp}}. Pour des jeux de données comportant de nombreux liens, l'execution de cette fontion
sera plus rapide que l'execution de \code{\link{proba_inclusion_indirecte}} (moins de lectures de  tables).
}
\examples{
poids_ind(UA=tbis_dataPop, dataStrate=tbis_dataStrate, l_ji=tbis_liens,  id_j="idA", id_i="idB", nomStrate="strate", population = "Nh", echantillon = "nh", controles=T)

}
\seealso{
\href{http://jms-insee.fr/jms2022s30_3/}{JMS 2022 : Utilisation des probabilités d’inclusion exactes pour le sondage indirect en population asymétrique}
}
