% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poids_mgpp.R
\name{poids_mgpp}
\alias{poids_mgpp}
\title{Methode géneralisée de Partage des poids}
\usage{
poids_mgpp(UA, sA, l_ji, id_j, w_j, id_i, aux_j, controles = T, sorties = T)
}
\arguments{
\item{UA}{base de sondage (variables id_j + aux_j)}

\item{sA}{echantillon tire (variables id_j, w_j)}

\item{l_ji}{table des liens (variables id_j, id_i) : pour le moment on n'autorise pas que j soit lie a plusieurs i}

\item{id_j}{identifiant au niveau UA}

\item{w_j}{poids de sondage des unites de sA}

\item{id_i}{identifiant au niveau UB}

\item{aux_j}{variable auxiliaire au niveau UA}

\item{controles}{effectuer des controles sur les tables (default=T)}

\item{sorties}{ecrit un petit resume a la fin (a ameliorer)}
}
\value{
2 tables sont produites dans une liste en sortie

La table w_i comporte 5 variables :

* w_i_cl : poids MGPP classiques

* w_i_aux : poids MGPP avec liens ponderes par la variable aux_j

* w_i_cl_exh : vaut w_i_cl, sauf si au moins une unite liee appartient a l'exhaustif (w_i_cl_exh vaut alors 1)

* w_i_aux_exh : vaut w_i_aux, sauf si au moins une unite liee appartient a l'exhaustif (w_i_aux_exh vaut alors 1)

* exh : indicatrice d'appartenance a l'exhaustif : une unite de UB est dans l'exhaustif si au moins une unite qui lui est lie a w_j=1


la table teta servira pour calculer les variables zk (utiles pour des traitements type calcul de precision, calage...)
}
\description{
Methode géneralisée de Partage des poids
}
\examples{
poids_mgpp(UA=tbis_dataPop, sA=tbis_sA, l_ji=tbis_liens, id_j="idA", w_j="poids", id_i="idB", aux_j="CA")

}
\seealso{
\href{http://jms-insee.fr/jms2018s24_4/}{JMS 2018 : La gestion par partage des poids des changements de contour des entreprises dans l’Enquete Sectorielle Annuelle}
}
