% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_inclusion_indirecte.R
\name{proba_inclusion_indirecte}
\alias{proba_inclusion_indirecte}
\title{Probabilite inclusion indirecte}
\usage{
proba_inclusion_indirecte(
  liste_liees,
  dataPop = NULL,
  dataStrate = NULL,
  identifiant = "id",
  nomStrate = "idStrate",
  population = "Nh",
  echantillon = "nh"
)
}
\arguments{
\item{liste_liees}{vecteur avec identifiants des unités liée dans la table datapop}

\item{dataPop}{base de sondage}

\item{dataStrate}{dataframe decrivant les strates}

\item{identifiant}{variable identifiant dans dataPop}

\item{nomStrate}{variable strate dans dataPop et dans dataStrate}

\item{population}{variable indiquant le nombre d'unite dans la BDS dans la strate dans dataStrate}

\item{echantillon}{variable indiquant le nombre d'unite a tirer dans la strate dans dataStrate}
}
\value{
probabilite d'inclusion indirecte de l'unite lie
}
\description{
Cette fonction prend en entree un vecteur d'identifiants d'unités liées entre elles et fournit en retour la probabilite d'inclusion indirecte pour un SAS stratifie
}
\examples{

### calcul des probas d'inclusion d'une unité liée à id1,id6 et id12
proba_inclusion_indirecte(c("id1","id6","id12"),
dataPop = population,dataStrate =stTest,
identifiant = "id",
 nomStrate = "strate",
 population = "N",
 echantillon =  "n")
}
