% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stTest}
\alias{stTest}
\title{Table de strate coherente avec population, genere l'exemple}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
stTest
}
\description{
Table de strate coherente avec population, genere l'exemple
}
\keyword{datasets}
