% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbis_dataPop}
\alias{tbis_dataPop}
\title{Base de sondage generee pour l'exemple}
\format{
An object of class \code{data.frame} with 160 rows and 3 columns.
}
\usage{
tbis_dataPop
}
\description{
Base de sondage generee pour l'exemple
}
\keyword{datasets}
