% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbis_dataStrate}
\alias{tbis_dataStrate}
\title{Table de strate exemple}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
tbis_dataStrate
}
\description{
Table de strate exemple
}
\keyword{datasets}
