% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbis_liens}
\alias{tbis_liens}
\title{Table de liens entre unitées initiales (idA) et finales (idB) exemple}
\format{
An object of class \code{data.frame} with 160 rows and 2 columns.
}
\usage{
tbis_liens
}
\description{
Table de liens entre unitées initiales (idA) et finales (idB) exemple
}
\keyword{datasets}
