% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tbis_sA}
\alias{tbis_sA}
\title{Echantillon initial exemple}
\format{
An object of class \code{data.frame} with 20 rows and 4 columns.
}
\usage{
tbis_sA
}
\description{
Echantillon initial exemple
}
\keyword{datasets}
