% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poids_mgpp.R
\name{poids_mgpp}
\alias{poids_mgpp}
\title{Methode generalisee de Partage des poids}
\usage{
poids_mgpp(UA, sA, l_ji, id_j, w_j, id_i, aux_j, controles = T, sorties = T)
}
\arguments{
\item{UA}{base de sondage (variables id_j + aux_j)}

\item{sA}{echantillon tire (variables id_j, w_j)}

\item{l_ji}{table des liens (variables id_j, id_i) : pour le moment on n'autorise pas que j soit lie a plusieurs i}

\item{id_j}{identifiant au niveau UA}

\item{w_j}{poids de sondage des unites de sA}

\item{id_i}{identifiant au niveau UB}

\item{aux_j}{variable auxiliaire au niveau UA}

\item{controles}{effectuer des controles sur les tables (default=T)}

\item{sorties}{ecrit un petit resume a la fin (a ameliorer)}
}
\value{
2 tables sont produites dans une liste en sortie

La table w_i comporte 5 variables :

* w_i_cl : poids MGPP classiques

* w_i_aux : poids MGPP avec liens ponderes par la variable aux_j

* w_i_cl_exh : vaut w_i_cl, sauf si au moins une unite liee appartient a l'exhaustif (w_i_cl_exh vaut alors 1)

* w_i_aux_exh : vaut w_i_aux, sauf si au moins une unite liee appartient a l'exhaustif (w_i_aux_exh vaut alors 1)

* exh : indicatrice d'appartenance a l'exhaustif : une unite de UB est dans l'exhaustif si au moins une unite qui lui est lie a w_j=1


la table teta servira pour calculer les variables zk (utiles pour des traitements type calcul de precision, calage...)
}
\description{
Methode generalisee de Partage des poids
}
\examples{

# Exemples article JMS (pages 4 a 9)

# Exemple 3.1

bds_ul_31 <- data.frame(
  id_ul=c("a","b","c"),
  CA=c(1,5,10)
)

ech_ul_31 <- data.frame(
  id_ul=c("a","b","c"),
  w_ul=c(10,10,10)
)

liens_31 <- data.frame(
  id_ul=c("a","b","c"),
  id_EP=c("A'","B'","B'")
)

w_ex31 <- poids_mgpp(UA=bds_ul_31, sA=ech_ul_31, l_ji=liens_31, id_j="id_ul", w_j="w_ul", id_i="id_EP", aux_j="CA")

#' on constate que les EP ont un poids final de 10 quelque soit la methode (table w_ex31$w_i), et que les teta sont ceux indiques dans letude (table w_ex31$teta)


# Exemple 3.2

bds_ul_32 <- data.frame(
  id_ul=c("a","b","c"),
  CA=c(1,5,10)
)

ech_ul_32 <- data.frame(
  id_ul=c("a","b","c"),
  w_ul=c(10,1,1)
)

liens_32 <- data.frame(
  id_ul=c("a","b","c"),
  id_EP=c("A'","A'","A'")
)

w_ex32 <- poids_mgpp(UA=bds_ul_32, sA=ech_ul_32, l_ji=liens_32, id_j="id_ul", w_j="w_ul", id_i="id_EP", aux_j="CA")




# Exemple 3.3

bds_ul_33 <- data.frame(
  id_ul=c("a","b","c"),
  CA=c(1,5,10)
)

ech_ul_33 <- data.frame(
  id_ul=c("b","c"),
  w_ul=c(1,1)
)

liens_33 <- data.frame(
  id_ul=c("a","b","c"),
  id_EP=c("A'","A'","A'")
)

w_ex33 <- poids_mgpp(UA=bds_ul_33, sA=ech_ul_33, l_ji=liens_33, id_j="id_ul", w_j="w_ul", id_i="id_EP", aux_j="CA")



# Exemple 3.4

bds_ul_34 <- data.frame(
  id_ul=c("a","b","c"),
  CA=c(1,5,10)
)

ech_ul_34 <- data.frame(
  id_ul=c("a"),
  w_ul=c(10)
)

liens_34 <- data.frame(
  id_ul=c("a","b","c"),
  id_EP=c("A'","A'","A'")
)

w_ex34 <- poids_mgpp(UA=bds_ul_34, sA=ech_ul_34, l_ji=liens_34, id_j="id_ul", w_j="w_ul", id_i="id_EP", aux_j="CA")



# Exemple 3.5

bds_ul_35 <- data.frame(
  id_ul=c("a","b","c"),
  CA=c(1,5,10)
)

ech_ul_35 <- data.frame(
  id_ul=c("a","b","c"),
  w_ul=c(10,1,1)
)

liens_35 <- data.frame(
  id_ul=c("a","b","c"),
  id_EP=c("A'","A'","B'")
)

w_ex35 <- poids_mgpp(UA=bds_ul_35, sA=ech_ul_35, l_ji=liens_35, id_j="id_ul", w_j="w_ul", id_i="id_EP", aux_j="CA")


}
\seealso{
http://jms-insee.fr/jms2018s24_4/
}
