% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{Base de sondage generee pour l'exemple}
\format{
An object of class \code{data.frame} with 50 rows and 2 columns.
}
\usage{
population
}
\description{
Base de sondage generee pour l'exemple
}
\keyword{datasets}
