% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_double_indirecte.R
\name{proba_double_indirecte}
\alias{proba_double_indirecte}
\title{Calcul de la probabilité d'inclusion double}
\usage{
proba_double_indirecte(liste1, liste2, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Calcul de la probabilité d'inclusion double
}
