% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_double_indirecte.R
\name{proba_double_indirecte}
\alias{proba_double_indirecte}
\title{Probabilité d'inclusion double indirecte}
\usage{
proba_double_indirecte(liste1, liste2, ...)
}
\arguments{
\item{liste1}{vecteur des identifiants des unites de dataPop liées a la premiere unite}

\item{liste2}{vecteur des identifiants des unites de dataPop liées a la seconde unite}

\item{...}{le reste des arguments est identique a la fonction proba_inclusion_indirecte}
}
\value{

}
\description{
Probabilité d'inclusion double indirecte
}
\examples{
# calcul de la probabilité d'inclusion double des unités liées c("id1","id6","id12") et c("id2","id5")
proba_double_indirecte(c("id1","id6","id12"),c("id2","id5"),
dataPop = population,
dataStrate =stTest,
identifiant = "id",
nomStrate = "strate",
population = "N",
echantillon =  "n")

}
