% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_double_indirecte.R
\name{proba_double_indirecte}
\alias{proba_double_indirecte}
\title{Probabilité d'inclusion double indirecte}
\usage{
proba_double_indirecte(liste1, liste2, ...)
}
\arguments{
\item{liste1}{vecteur des identifiants des unites de dataPop liées a la premiere unite liée}

\item{liste2}{vecteur des identifiants des unites de dataPop liées a la seconde unite liée}

\item{...}{le reste des arguments est identique a la fonction \code{\link{proba_inclusion_indirecte}}}
}
\value{
probabilite d'inclusion double indirecte des deux unités liées
}
\description{
Cette fonction est proche de \code{\link{proba_inclusion_indirecte}}
mais permet de calculer une probabilité d'inclusion double entre deux unités finales.
}
\examples{
# calcul de la probabilité d'inclusion double des unités liées c("32","43","111") et c("51","103")
proba_double_indirecte(c("32","43","111"),c("51","103"),
                      dataPop = tbis_dataPop,
                      dataStrate =tbis_dataStrate,
                      identifiant = "idA",
                      nomStrate = "strate",
                      population = "Nh",
                      echantillon =  "nh")
}
\seealso{
\href{http://jms-insee.fr/jms2022s30_3/}{JMS 2022 : Utilisation des probabilités d’inclusion exactes pour le sondage indirect en population asymétrique}
}
