% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_inclusion_indirecte.R
\name{proba_inclusion_indirecte}
\alias{proba_inclusion_indirecte}
\title{Probabilité d'inclusion simple indirecte}
\usage{
proba_inclusion_indirecte(
  liste_liees,
  dataPop = NULL,
  dataStrate = NULL,
  identifiant = "id",
  nomStrate = "idStrate",
  population = "Nh",
  echantillon = "nh"
)
}
\arguments{
\item{liste_liees}{vecteur avec identifiants des unités liée dans la table datapop}

\item{dataPop}{base de sondage}

\item{dataStrate}{dataframe decrivant les strates}

\item{identifiant}{variable identifiant dans dataPop}

\item{nomStrate}{variable strate dans dataPop et dans dataStrate}

\item{population}{variable indiquant le nombre d'unite dans la BDS dans la strate dans dataStrate}

\item{echantillon}{variable indiquant le nombre d'unite a tirer dans la strate dans dataStrate}
}
\value{
probabilite d'inclusion indirecte de l'unite liée
}
\description{
Cette fonction prend en entree un vecteur d'identifiants d'unités liées entre elles et fournit en retour la probabilite d'inclusion indirecte pour un SAS stratifie.

La fonction \code{\link{proba_double_indirecte}} permet de calculer des probabilités d'inclusion double.

Pour des jeux de données comportant de nombreux liens, l'execution de \code{\link{proba_indirecte_fast}} devrait être plus rapide (moins de lectures de  tables).
}
\examples{

#exemple pour idB==21


proba_inclusion_indirecte(c("32","43","111"),
                         dataPop = tbis_dataPop,
                         dataStrate =tbis_dataStrate,
                         identifiant = "idA",
                         nomStrate = "strate",
                         population = "Nh",
                         echantillon =  "nh")

}
\seealso{
\href{http://jms-insee.fr/jms2022s30_3/}{JMS 2022 : Utilisation des probabilités d’inclusion exactes pour le sondage indirect en population asymétrique}
}
